+++ healthchecks/src/test/java/uk/gov/justice/services/healthcheck/healthchecks/EventStoreHealthcheckTest.java
@@ -1,0 +1,84 @@
+package uk.gov.justice.services.healthcheck.healthchecks;
+
+import static java.util.Optional.of;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+import static uk.gov.justice.services.healthcheck.healthchecks.EventStoreHealthcheck.EVENT_STORE_TABLE_NAMES;
+
+import uk.gov.justice.services.eventsourcing.source.core.EventStoreDataSourceProvider;
+import uk.gov.justice.services.healthcheck.api.HealthcheckResult;
+import uk.gov.justice.services.healthcheck.database.TableChecker;
+
+import java.sql.SQLException;
+import java.util.Optional;
+
+import javax.inject.Inject;
+import javax.sql.DataSource;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
+import org.slf4j.Logger;
+
+@RunWith(MockitoJUnitRunner.class)
+public class EventStoreHealthcheckTest {
+
+    @Mock
+    private EventStoreDataSourceProvider eventStoreDataSourceProvider;
+
+    @Mock
+    private TableChecker tableChecker;
+
+    @Mock
+    private Logger logger;
+
+    @InjectMocks
+    private EventStoreHealthcheck eventStoreHealthcheck;
+
+    @Test
+    public void shouldReturnCorrectHealthcheckName() throws Exception {
+
+        assertThat(eventStoreHealthcheck.getHealthcheckName(), is("event-store-healthcheck"));
+    }
+
+    @Test
+    public void shouldReturnCorrectHealthcheckDescription() throws Exception {
+
+        assertThat(eventStoreHealthcheck.healthcheckDescription(), is("Checks connectivity to the eventstore database and that all framework tables are available"));
+    }
+
+    @Test
+    public void shouldGetListOfExpectedTablesFromEventStoreAsHealthcheck() throws Exception {
+
+        final DataSource eventStoreDataSource = mock(DataSource.class);
+        final HealthcheckResult healthcheckResult = mock(HealthcheckResult.class);
+
+        when(eventStoreDataSourceProvider.getDefaultDataSource()).thenReturn(eventStoreDataSource);
+        when(tableChecker.checkTables(EVENT_STORE_TABLE_NAMES, eventStoreDataSource)).thenReturn(healthcheckResult);
+
+        assertThat(eventStoreHealthcheck.runHealthcheck(), is(healthcheckResult));
+    }
+
+    @Test
+    public void shouldReturnHealthcheckFailureIfAccessingTheEventStoreThrowsSqlException() throws Exception {
+
+        final SQLException sqlException = new SQLException("Oops");
+        final DataSource eventStoreDataSource = mock(DataSource.class);
+
+        when(eventStoreDataSourceProvider.getDefaultDataSource()).thenReturn(eventStoreDataSource);
+        when(tableChecker.checkTables(EVENT_STORE_TABLE_NAMES, eventStoreDataSource)).thenThrow(sqlException);
+
+        final HealthcheckResult healthcheckResult = eventStoreHealthcheck.runHealthcheck();
+
+        assertThat(healthcheckResult.isPassed(), is(false));
+        assertThat(healthcheckResult.getErrorMessage().isPresent(), is(true));
+        assertThat(healthcheckResult.getErrorMessage(), is(of("Exception thrown accessing eventstore database. java.sql.SQLException: Oops")));
+
+        verify(logger).error("Healthcheck for eventstore database failed.", sqlException);
+    }
+}